+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.0
+
+- remove PhysBone action
+- replace JSON operations with Newtonsoft
+- more visible errors
+- hide errors on save actions
+- added "show" button to paths
+- fix font color
+- re-order steps
+
 # alpha2
 
 - switch material and remove game object custom actions working